+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **Feature:** Added default UI support for managing custom permissions in case `requestCustomPermissions` hasn't been implemented in Host App.
 - **Feature:** Added `MiniAppHasNoPublishedVersionException` and `MiniAppNotFoundException` exception types to `MiniApp.create` and `MiniApp.fetchInfo`.
 - **Feature:** Mini App can call media execution play/pause programmatically.
+- **Change:** Added the default implementation for external link handler. Using [custom tab](https://developers.google.com/web/android/custom-tabs).
 
 ### 2.4.0 (2020-10-30)
 **SDK**